feat: support egress and inspection rule types (#4)
**Issue #, if available:**

## Description of changes:

`Egress` and `Inspection` traffic can be routed differently. For this
reason, the ability to define the type of traffic has been added.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.