fix: implement handshake rule for non 443 ports with tls requirements (#11)
**Issue #, if available:**

## Description of changes:

When tis versions are defined and the port was not 443. The handshake
rule was not generated. With this fix it's get generated again.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [ ] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.