feat: generate rules for dns servers
In some cases you might want to sync Route53 resolvers with on-premise. For this you need to have 2 rules a TCP and a UDP rule and it needs to be bidirectional. For this reason I implemented it on the protocol level. I also made a simpler implementation for controlling when something should be bidirectional.