feat: generate rules for dns servers (#14)
**Issue #, if available:**

## Description of changes:

In some cases you might want to sync Route53 resolvers with on-premise.
For this you need to have 2 rules a TCP and a UDP rule and it needs to
be bidirectional. For this reason I implemented it on the protocol
level. I also made a simpler implementation for controlling when
something should be bidirectional.


**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [ ] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.