fix: tls rules without endpoints (#23)
**Issue #, if available:**

## Description of changes:

In some cases you do not want to match on the `SNI` (Server Name
Indicator). But you do want to enforce the usage of TLS on the
connection. In these cases we should not try to add the SNI option in
the suricata rule.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [ ] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.